requirements.txt
chromadb>=0.4.0
langchain>=0.1.0
openai>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
openpyxl>=3.1.0
